Pull segments from front of stack to evenly apply simplification.
Stack-based recursion is depth-first, so the initial "half" of the
points array is simplified first. Depending on the specified maximum
number of points, this can produce results where the first "half" of
points are accurately simplified, but the second "half" has been
reduced to a straight (or mostly straight) line.